+++ Sources/ViewModel/ChannelSettings/SBUOpenChannelSettingsViewModel.swift
@@ -43,6 +43,7 @@ open class SBUOpenChannelSettingsViewModel: SBUBaseChannelSettingsViewModel {
         delegate: SBUOpenChannelSettingsViewModelDelegate? = nil
     ) {
         super.init()
+        self.commonDelegate = delegate
         
         self.delegate = delegate
         self.baseDelegates.addDelegate(delegate, type: .uikit)